scalper: resolve the traded contract from the exchange, not from date math
instruments.py knew a /MES's economics from a hardcoded table and its front month
from date arithmetic (quarterly codes + an 8-day roll heuristic). Both were
conventions we typed in, and when they are wrong they fail silently: a stale front
month still streams, still fires the detector, still fills the paper broker — on a
book nobody is trading.

Adds tastytrade GET /instruments/futures (FuturesRequest / FutureContract /
FuturesResponse in endpoints/tastytrade.py), which returns every listed contract
with its active-month flag, dxFeed streamer-symbol, notional-multiplier, tick-size
and tradeability flags.

contract.py::resolve_contract runs it at daemon startup, before anything keys off
the symbol:

- a bare root (--symbols /MES, now the working default) becomes the exchange's own
  active-month contract, and the plan path / broker economics / subscribe list all
  follow from that;
- a dated symbol is honored (the plan named it on purpose) but checked — warns when
  it is no longer the front month or is closing-only;
- the exchange's notional-multiplier / tick-size override the registry, loudly: a
  disagreement means a constant in instruments.py is wrong, and every prior paper
  session on that root priced its P&L with it;
- any failure (no creds, no network, --demo-setup) degrades to the date math with
  verified=False, and the banner prints UNVERIFIED (date math). Prep works offline.

The registry keeps what the API has no opinion on: the detector's Geometry bands
(calibration) and the cash-index reference (our level-source decision).

Verified against the live API: /MES $5, /MNQ $2, /ES $50, /NQ $20, all 0.25 tick,
zero reconcile warnings; front month resolves to U26 for all four. `--basis
--symbols /MES` end-to-end prints /MESU26:XCME − SPX = +44.41.

Also reflows one signature in dxlink_stream_client.py (ruff format drift, no
behavior change).

284 tests green; ruff + ty clean.

+++ packages/trading-scalper/src/trading_scalper/paper.py
@@ -68,7 +68,7 @@ class _Resting:
 class PaperBroker:
     """A ``BrokerExecution`` backed by an in-memory matching engine."""
 
-    def __init__(self, multiplier: int = 100, tick: float = 0.25) -> None:
+    def __init__(self, multiplier: float = 100, tick: float = 0.25) -> None:
         self._ids = count(1)
         self._ledger = Ledger(multiplier=multiplier)
         self._tick = tick
